requirements.txt
requests>=2.31.0
customtkinter>=5.2.2
Pillow>=10.0.0
dnspython>=2.6.1
screeninfo>=0.8
